fix code block copy button
